get todo list by id integration test

+++ server/models/todoList.js
@@ -3,7 +3,10 @@ const {todoList} = require('../database/dbSetup')
 module.exports = {
     clearTodoLists: async () => {
         try {
-            await todoList.collection.drop()
+            const collectionList = await todoList.db.db.listCollections({name: todoList.collection.name}).toArray()
+            if (collectionList.length !== 0) {
+                await todoList.collection.drop()
+            }          
         } catch (error) {
             console.error(error)
         }
